refactor: extract run_query into core module
