+++ src/gcp.rs
@@ -63,68 +63,8 @@ impl DatastoreService for DatastoreEmulator {
         &self,
         request: Request<RunQueryRequest>,
     ) -> Result<Response<RunQueryResponse>, Status> {
-        tracing::debug!("Received RunQueryRequest: {:?}", request);
-        let start = Instant::now();
-        let req = request.into_inner();
-        let storage = self.storage.read().await;
-        // Extract the query from the request
-        let query_obj = match req.query_type {
-            Some(crate::google::datastore::v1::run_query_request::QueryType::Query(query)) => query,
-            _ => {
-                // implement other query types if needed
-                return Err(Status::invalid_argument("Missing or invalid query"));
-            }
-        };
-        let kind_name = query_obj
-            .kind
-            .first()
-            .map(|k| k.name.clone())
-            .ok_or_else(|| Status::invalid_argument("Query must specify a kind"))?;
-        let batch = storage.get_entities(
-            req.project_id.clone(),
-            kind_name,
-            query_obj.filter.clone(),
-            query_obj.limit.as_ref().map(|v| v.value),
-            query_obj.start_cursor.clone(),
-            query_obj.projection.clone(),
-            query_obj.distinct_on.clone(),
-            query_obj.order.clone(),
-        );
-        let mut fields = HashMap::new();
-
-        let amount_results = batch.entity_results.len() as i64;
-
-        fields.insert(
-            "Some key".to_string(),
-            ValueProps {
-                kind: Some(Kind::StringValue("Some value".to_string())),
-            },
-        );
-        let debug_stats = Struct {
-            fields: fields.clone(),
-        };
-
-        let execution_duration = start.elapsed();
-
-        Ok(Response::new(RunQueryResponse {
-            transaction: vec![],
-            query: Some(query_obj),
-            batch: Some(batch),
-            // todo: create real metrics
-            explain_metrics: Some(ExplainMetrics {
-                plan_summary: Some(PlanSummary {
-                    indexes_used: vec![Struct {
-                        fields: fields.clone(),
-                    }],
-                }),
-                execution_stats: Some(ExecutionStats {
-                    results_returned: amount_results,
-                    execution_duration: Some(to_prost_duration(execution_duration)),
-                    read_operations: 10,
-                    debug_stats: Some(debug_stats),
-                }),
-            }),
-        }))
+        let resp = crate::core::run_query(&self.storage, request.into_inner()).await?;
+        Ok(Response::new(resp))
     }
 
     async fn commit(
